+++ src/Mark.ts
@@ -1,54 +1,37 @@
 'use strict';
 
 import * as vscode from 'vscode';
-import * as _ from 'lodash';
 
 import { Helper } from './Helper';
 import { File } from './File';
 
-import { debLog, debIn, debOut, debBackFrom } from './DebLog';
-
-const className = 'Mark';
-function ind(methodName: string, text = '') {
-  debIn(className, methodName, text);
-}
-function out(methodName: string, text = '') {
-  debOut(className, methodName, text);
-}
-function log(methodName: string, text = ''): number {
-  return debLog(className, methodName, text);
-}
-function backFrom(count: number, methodName: string, text = '') {
-  debBackFrom(count, className, methodName, text);
-}
-
 export class Mark {
   private _isDirty: boolean;
   private _parent: File;
-  private _lineNumber: number;
-  private _quickPickItem: vscode.QuickPickItem;
-  private _dirtyLineNumber: number;
-  private _dirtyQuickPickItem: vscode.QuickPickItem;
+  private _lineNumber: number | undefined;
+  private _quickPickItem: vscode.QuickPickItem | undefined;
+  private _dirtyLineNumber: number | undefined;
+  private _dirtyQuickPickItem: vscode.QuickPickItem | undefined;
 
-  public get quickPickItem(): vscode.QuickPickItem {
+  public get quickPickItem(): vscode.QuickPickItem | undefined {
     if (this._isDirty) {
       return this._dirtyQuickPickItem;
     }
     return this._quickPickItem;
   }
 
-  public get lineNumberForPersist(): number {
+  public get lineNumberForPersist(): number | undefined {
     return this._lineNumber;
   }
 
-  public get lineNumber(): number {
+  public get lineNumber(): number | undefined {
     if (this._isDirty) {
       return this._dirtyLineNumber;
     }
     return this._lineNumber;
   }
 
-  public set lineNumber(lineNumber: number) {
+  public set lineNumber(lineNumber: number | undefined) {
     this._isDirty = true;
     this.setLineNumber(lineNumber);
   }
@@ -59,14 +42,13 @@ export class Mark {
     this.setLineNumber(lineNumber);
   }
 
-  private setLineNumber(lineNumber: number) {
+  private setLineNumber(lineNumber: number | undefined) {
     if (this._isDirty) {
       this._dirtyLineNumber = lineNumber;
     } else {
       this._lineNumber = lineNumber;
     }
 
-    let debCount = log('lineNumber', 'call Helper.getQuickPickItem(...)');
     Helper.getQuickPickItem(this._parent.filepath, lineNumber)
       .then(value => {
         this._quickPickItem = value;