+++ src/Ring.ts
@@ -1,8 +1,8 @@
-"use strict";
+'use strict';
 
 // not a 'real' ring class - only stuff I need
 export class Ring<T> extends Array<T> {
-  private _current: number;
+  private _current: number = 0;
 
   public get current(): T {
     return this[this._current];
@@ -24,15 +24,15 @@ export class Ring<T> extends Array<T> {
     return this.current;
   }
 
-  public insert(...items: T[]): number {
+  public insert(...items: T[]): T {
     super.splice(this._current, 0, ...items);
     this._current += items.length;
-    return this._current;
+    return this.current;
   }
 
   push(...items: T[]): number {
     super.push(...items);
     this._current = super.length - 1;
-    return this._current;
+    return items.length;
   }
 }